fix: fix hooks error at useErrorCatch
- remove useErrorCatch
- add Errorboundary component
