+++ package.json
@@ -47,3 +47,4 @@
     "workerDirectory": "public"
   }
 }
+
